Change skin mnist into skin condition task
